Move merge settings method to Node
