+++ server/src/main/java/org/elasticsearch/plugins/PluginsService.java
@@ -236,31 +236,6 @@ public class PluginsService implements ReportingService<PluginsAndModules> {
         return plugins.stream().collect(Collectors.toMap(p -> p.info().getName(), LoadedPlugin::instance));
     }
 
-    public static Settings mergedPluginSettings(Map<String, Plugin> pluginMap) {
-        Map<String, String> foundSettings = new HashMap<>();
-        final Settings.Builder builder = Settings.builder();
-        for (Map.Entry<String, Plugin> entry : pluginMap.entrySet()) {
-            Settings settings = entry.getValue().additionalSettings();
-            for (String setting : settings.keySet()) {
-                String oldPlugin = foundSettings.put(setting, entry.getKey());
-                if (oldPlugin != null) {
-                    throw new IllegalArgumentException(
-                        "Cannot have additional setting ["
-                            + setting
-                            + "] "
-                            + "in plugin ["
-                            + entry.getKey()
-                            + "], already added in plugin ["
-                            + oldPlugin
-                            + "]"
-                    );
-                }
-            }
-            builder.put(settings);
-        }
-        return builder.build();
-    }
-
     /**
      * Get information about plugins and modules
      */
