+++ server/src/test/java/org/elasticsearch/plugins/PluginsServiceTests.java
@@ -15,6 +15,7 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.env.TestEnvironment;
 import org.elasticsearch.index.IndexModule;
+import org.elasticsearch.node.Node;
 import org.elasticsearch.test.ESTestCase;
 
 import java.io.IOException;
@@ -91,7 +92,7 @@ public class PluginsServiceTests extends ESTestCase {
             .put(IndexModule.INDEX_STORE_TYPE_SETTING.getKey(), IndexModule.Type.NIOFS.getSettingsKey())
             .build();
         Map<String, Plugin> pluginMap = Map.of(AdditionalSettingsPlugin1.class.getName(), new AdditionalSettingsPlugin1());
-        Settings newSettings = Settings.builder().put(PluginsService.mergedPluginSettings(pluginMap)).put(settings).build();
+        Settings newSettings = Settings.builder().put(Node.mergedPluginSettings(pluginMap)).put(settings).build();
         assertEquals("test", newSettings.get("my.setting")); // previous settings still exist
         assertEquals("1", newSettings.get("foo.bar")); // added setting exists
         // does not override pre existing settings
@@ -105,7 +106,7 @@ public class PluginsServiceTests extends ESTestCase {
             AdditionalSettingsPlugin2.class.getName(),
             new AdditionalSettingsPlugin2()
         );
-        IllegalArgumentException e = expectThrows(IllegalArgumentException.class, () -> PluginsService.mergedPluginSettings(pluginMap));
+        IllegalArgumentException e = expectThrows(IllegalArgumentException.class, () -> Node.mergedPluginSettings(pluginMap));
         String msg = e.getMessage();
         assertTrue(msg, msg.contains("Cannot have additional setting [foo.bar]"));
         assertTrue(msg, msg.contains("plugin [" + AdditionalSettingsPlugin1.class.getName()));
